+++ tests/tests_pytorch/checkpointing/test_model_checkpoint.py
@@ -556,6 +556,15 @@ def test_model_checkpoint_link_checkpoint(tmp_path):
     assert os.path.islink(folder_link)
     assert os.path.realpath(folder_link) == str(new_folder)
 
+    # simulate permission error on Windows (creation of symbolic links requires privileges)
+    file = tmp_path / "win_file"
+    file.touch()
+    link = tmp_path / "win_link"
+    with mock.patch("lightning.pytorch.callbacks.model_checkpoint.os.symlink", Mock(side_effect=OSError)):
+        ModelCheckpoint._link_checkpoint(trainer, filepath=str(file), linkpath=str(link))
+    assert not os.path.islink(link)
+    assert os.path.isfile(link)  # fall back to copying instead of linking
+
 
 def test_invalid_top_k(tmpdir):
     """Make sure that a MisconfigurationException is raised for a negative save_top_k argument."""